修复无网络的状况的问题

+++ Star/Main/Model/HomeModel.swift
@@ -31,7 +31,7 @@ class HomeUserEntity: BaseModel {
 }
 class InviteEntity: BaseModel {
     @objc var inviteCount : String?
-    @objc var invitePower : Double = 0
+    @objc var invitePower : Int = 0
     @objc var inviteCode : String?
     @objc var nickname : String?
 }
